fix(CI): bugs using workflows variables and schema serialization
