+++ dessia_common/serialization.py
@@ -20,7 +20,7 @@ from dessia_common.abstract import CoreDessiaObject
 from dessia_common.typings import InstanceOf, JsonSerializable
 from dessia_common.graph import explore_tree_from_leaves  # , cut_tree_final_branches
 from dessia_common.breakdown import get_in_object_from_path
-from dessia_common.schemas.core import TYPING_EQUIVALENCES, is_typing
+from dessia_common.schemas.core import TYPING_EQUIVALENCES, is_typing, serialize_typing
 
 fullargsspec_cache = {}
 
@@ -130,7 +130,7 @@ def serialize(value):
     elif isinstance(value, float64):
         serialized_value = float(value)
     elif isinstance(value, type) or is_typing(value):
-        return dcty.serialize_typing(value)
+        return serialize_typing(value)
     elif hasattr(value, 'to_dict'):
         to_dict_method = getattr(value, 'to_dict', None)
         if callable(to_dict_method):
@@ -169,7 +169,7 @@ def serialize_with_pointers(value, memo=None, path='#', id_method=True, id_memo=
     elif isinstance(value, type):
         if value in memo:
             return {'$ref': memo[value]}, memo
-        serialized = dcty.serialize_typing(value)
+        serialized = serialize_typing(value)
         # memo[value] = path
 
     # Regular object
