+++ dessia_common/workflow/blocks.py
@@ -6,12 +6,11 @@ import inspect
 import warnings
 
 from zipfile import ZipFile
-from typing import List, Type, Any, Dict, Tuple, get_type_hints
+from typing import List, Type, Any, Dict, Tuple, get_type_hints, TypeVar
 
 import itertools
 from dessia_common.core import DessiaFilter, FiltersList, type_from_annotation, DessiaObject
-from dessia_common.schemas.core import split_argspecs
-from dessia_common.utils.docstrings import parse_docstring, EMPTY_PARSED_ATTRIBUTE
+from dessia_common.schemas.core import split_argspecs, parse_docstring, EMPTY_PARSED_ATTRIBUTE
 from dessia_common.displays import DisplaySetting, DisplayObject
 from dessia_common.errors import UntypedArgumentError
 from dessia_common.typings import JsonSerializable, MethodType, ClassMethodType
@@ -23,6 +22,9 @@ from dessia_common.exports import ExportFormat
 from dessia_common.workflow.core import Block, Variable, TypedVariable, TypedVariableWithDefaultValue,\
     set_block_variable_names_from_dict, Workflow
 from dessia_common.workflow.utils import ToScriptElement
+
+
+T = TypeVar("T")
 
 
 def set_inputs_from_function(method, inputs=None):
@@ -331,7 +333,7 @@ class Sequence(Block):
     def __init__(self, number_arguments: int, name: str = '', position: Tuple[float, float] = None):
         self.number_arguments = number_arguments
         inputs = [Variable(name=f"Sequence element {i}") for i in range(self.number_arguments)]
-        outputs = [TypedVariable(type_=list, name='sequence')]
+        outputs = [TypedVariable(type_=list[T], name='sequence')]
         Block.__init__(self, inputs, outputs, name=name, position=position)
 
     def equivalent_hash(self):
@@ -385,7 +387,7 @@ class Concatenate(Block):
     def __init__(self, number_arguments: int = 2, name: str = '', position: Tuple[float, float] = None):
         self.number_arguments = number_arguments
         inputs = [Variable(name=f"Sequence element {i}") for i in range(self.number_arguments)]
-        outputs = [TypedVariable(type_=list, name='sequence')]
+        outputs = [TypedVariable(type_=list[T], name='sequence')]
         Block.__init__(self, inputs, outputs, name=name, position=position)
 
     def equivalent_hash(self):
