+++ tests/utils/types.py
@@ -2,11 +2,11 @@ import unittest
 
 from parameterized import parameterized
 
-from dessia_common.utils.types import deserialize_typing, serialize_typing, MethodType, ClassMethodType, InstanceOf,\
-    is_jsonable
+from dessia_common.utils.types import deserialize_typing, MethodType, ClassMethodType, InstanceOf, is_jsonable
 from dessia_common.files import BinaryFile, StringFile
 from typing import List, Tuple, Type, Dict
 from dessia_common.forms import StandaloneObject
+from dessia_common.schemas.core import serialize_typing
 
 
 class TestTypingSerializationValid(unittest.TestCase):
